requirements.txt
Flask==3.1.2
Flask-SocketIO==5.5.1
pypdf==6.1.1
requests==2.32.5
Pillow==11.3.0
pytest==8.4.2
psutil==7.1.0
reportlab==4.4.4
pdfminer.six==20250506
pytesseract==0.3.13
pdfplumber==0.11.7
unstructured==0.18.15
unstructured[pdf]==0.18.15
pi_heif==1.1.1
